inline file_service.js into start_dev_server.js
